✨ Replace React About page with templated page
This commit adds support for templated pages at the top level. Pages are
put in the `src/content/pages` directory. They do not require all of the
same arguments as a Post or Project, and the mdx is read and injected
into a Page template. This commit also removes the old about page.