✨ Better footnotes
In #216 I removed remarkGfm and was opting to rely on
gatsby-transformer-remark for footnote support. As it turns out, there
that plugin and gatsby-plugin-mdx don't seem to play nicely, and it
broke footnotes (they are rendered now as plaintext 😬)

As part of reverting that commit, I also added a new custom rehype
plugin that will improve the footnote links.